backend/api/routers/streaming_router.py
"""
Streaming Router for MasterX Quantum Intelligence Platform

Real-time streaming API that provides server-sent events and streaming
capabilities for live learning interactions and real-time updates.

🌊 STREAMING CAPABILITIES:
- Server-sent events for real-time updates
- Learning progress streaming
- Live chat streaming
- Real-time notifications
- Performance metrics streaming
- Event-driven updates

Author: MasterX AI Team - API Integration Division
Version: 1.0 - Phase 12 Enhanced Backend APIs Integration
"""

import asyncio
import json
import logging
from datetime import datetime
from typing import Dict, Any, List, Optional, AsyncGenerator
from fastapi import APIRouter, HTTPException, Depends
from fastapi.responses import StreamingResponse
from sse_starlette.sse import EventSourceResponse

from ..models import (
    StreamingEvent, StreamingEventType, UserProfile, BaseResponse
)
from ..auth import get_current_user, require_permission
from ..utils import APIResponseHandler

# Configure logging
logger = logging.getLogger(__name__)

# Create router
router = APIRouter()

# ============================================================================
# STREAMING SERVICE
# ============================================================================

class StreamingService:
    """Streaming service for real-time updates"""
    
    def __init__(self):
        self.active_streams = {}
        self.event_queue = asyncio.Queue()
        self.response_handler = APIResponseHandler()
        logger.info("🌊 Streaming Service initialized")
    
    async def create_event_stream(self, user_id: str, event_types: List[str]) -> AsyncGenerator[str, None]:
        """Create event stream for user"""
        
        stream_id = f"stream_{user_id}_{int(datetime.now().timestamp())}"
        self.active_streams[stream_id] = {
            "user_id": user_id,
            "event_types": event_types,
            "created_at": datetime.now()
        }
        
        try:
            # Send initial connection event
            yield f"data: {json.dumps({'type': 'connection', 'message': 'Stream connected', 'timestamp': datetime.now().isoformat()})}\n\n"
            
            # Stream events
            while stream_id in self.active_streams:
                try:
                    # Wait for events with timeout
                    event = await asyncio.wait_for(self.event_queue.get(), timeout=30.0)
                    
                    # Check if event is for this user and type
                    if (event.user_id == user_id and 
                        event.event_type.value in event_types):
                        
                        event_data = {
                            "type": event.event_type.value,
                            "data": event.data,
                            "timestamp": event.timestamp.isoformat()
                        }
                        
                        yield f"data: {json.dumps(event_data)}\n\n"
                
                except asyncio.TimeoutError:
                    # Send keepalive
                    yield f"data: {json.dumps({'type': 'keepalive', 'timestamp': datetime.now().isoformat()})}\n\n"
                
                except Exception as e:
                    logger.error(f"Streaming error: {e}")
                    yield f"data: {json.dumps({'type': 'error', 'message': str(e)})}\n\n"
                    break
        
        finally:
            # Cleanup
            if stream_id in self.active_streams:
                del self.active_streams[stream_id]
    
    async def send_event(self, event: StreamingEvent):
        """Send event to all relevant streams"""
        await self.event_queue.put(event)
    
    async def stream_learning_progress(self, user_id: str) -> AsyncGenerator[str, None]:
        """Stream learning progress updates"""
        
        try:
            # Send initial progress
            initial_progress = {
                "overall_progress": 65.5,
                "current_goal": "Learn Python Basics",
                "today_study_time": 45,
                "streak": 7
            }
            
            yield f"data: {json.dumps({'type': 'initial_progress', 'data': initial_progress})}\n\n"
            
            # Simulate real-time progress updates
            for i in range(10):
                await asyncio.sleep(5)  # Update every 5 seconds
                
                progress_update = {
                    "study_time_increment": 5,
                    "new_achievement": f"Achievement {i+1}" if i % 3 == 0 else None,
                    "skill_improvement": f"Skill {i+1}" if i % 2 == 0 else None,
                    "timestamp": datetime.now().isoformat()
                }
                
                yield f"data: {json.dumps({'type': 'progress_update', 'data': progress_update})}\n\n"
        
        except Exception as e:
            yield f"data: {json.dumps({'type': 'error', 'message': str(e)})}\n\n"
    
    async def stream_notifications(self, user_id: str) -> AsyncGenerator[str, None]:
        """Stream real-time notifications"""
        
        try:
            # Send welcome notification
            welcome = {
                "title": "Welcome to MasterX",
                "message": "Your learning journey continues!",
                "type": "info",
                "timestamp": datetime.now().isoformat()
            }
            
            yield f"data: {json.dumps({'type': 'notification', 'data': welcome})}\n\n"
            
            # Simulate periodic notifications
            notifications = [
                {"title": "Study Reminder", "message": "Time for your daily study session!", "type": "reminder"},
                {"title": "Achievement Unlocked", "message": "You've completed 5 lessons this week!", "type": "achievement"},
                {"title": "New Content", "message": "New Python exercises are available", "type": "content"},
                {"title": "Progress Update", "message": "You're 80% complete with your current goal", "type": "progress"}
            ]
            
            for i, notification in enumerate(notifications):
                await asyncio.sleep(10)  # Send every 10 seconds
                notification["timestamp"] = datetime.now().isoformat()
                yield f"data: {json.dumps({'type': 'notification', 'data': notification})}\n\n"
        
        except Exception as e:
            yield f"data: {json.dumps({'type': 'error', 'message': str(e)})}\n\n"

# Initialize service
streaming_service = StreamingService()

@router.get("/events")
async def stream_events(
    event_types: str = "chat_message,learning_update,notification",
    current_user: UserProfile = Depends(require_permission("streaming:read"))
):
    """Stream real-time events"""
    
    try:
        event_type_list = [t.strip() for t in event_types.split(",")]
        
        return EventSourceResponse(
            streaming_service.create_event_stream(current_user.user_id, event_type_list),
            media_type="text/event-stream"
        )
    except Exception as e:
        logger.error(f"Stream events error: {e}")
        raise HTTPException(status_code=500, detail="Failed to create event stream")

@router.get("/progress")
async def stream_progress(
    current_user: UserProfile = Depends(require_permission("streaming:read"))
):
    """Stream learning progress updates"""
    
    try:
        return EventSourceResponse(
            streaming_service.stream_learning_progress(current_user.user_id),
            media_type="text/event-stream"
        )
    except Exception as e:
        logger.error(f"Stream progress error: {e}")
        raise HTTPException(status_code=500, detail="Failed to stream progress")

@router.get("/notifications")
async def stream_notifications(
    current_user: UserProfile = Depends(require_permission("streaming:read"))
):
    """Stream real-time notifications"""
    
    try:
        return EventSourceResponse(
            streaming_service.stream_notifications(current_user.user_id),
            media_type="text/event-stream"
        )
    except Exception as e:
        logger.error(f"Stream notifications error: {e}")
        raise HTTPException(status_code=500, detail="Failed to stream notifications")

@router.post("/send-event")
async def send_event(
    event_data: Dict[str, Any],
    current_user: UserProfile = Depends(require_permission("streaming:write"))
):
    """Send a custom event to streams"""
    
    try:
        event = StreamingEvent(
            event_type=StreamingEventType(event_data["event_type"]),
            user_id=event_data.get("user_id", current_user.user_id),
            data=event_data["data"]
        )
        
        await streaming_service.send_event(event)
        
        return BaseResponse(message="Event sent successfully")
    except Exception as e:
        logger.error(f"Send event error: {e}")
        raise HTTPException(status_code=500, detail="Failed to send event")
