backend/requirements.txt
fastapi>=0.115.7
uvicorn>=0.26.0
pydantic>=2.10.5
python-multipart>=0.0.19
jinja2>=3.1.4
aiofiles>=24.1.0 
websockets>=14.0
motor>=3.7.0
pymongo>=4.10.1
bcrypt>=4.2.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
aioredis>=2.0.1
celery>=5.4.0
redis>=5.2.1
requests>=2.32.3
httpx>=0.28.1
python-dotenv>=1.0.1
email-validator>=2.2.0
pillow>=11.0.0
matplotlib>=3.9.3
seaborn>=0.13.2
plotly>=5.24.1
pandas>=2.2.3
numpy>=2.2.1
scipy>=1.14.1
scikit-learn>=1.6.0
transformers>=4.46.3
torch>=2.2.0
torchvision>=0.17.0
openai>=1.57.2
anthropic>=0.42.0
google-generativeai>=0.8.3
cohere>=5.15.0
groq>=0.13.0
langchain>=0.3.14
langchain-openai>=0.2.14
langchain-anthropic>=0.2.5
langchain-google-genai>=2.0.8
langchain-groq>=0.2.1
langchain-community>=0.3.14
langchain-core>=0.3.28
langchain-text-splitters>=0.3.2
chromadb>=0.6.4
faiss-cpu>=1.9.0.post1
sentence-transformers>=3.3.1
tiktoken>=0.8.0
pypdf>=5.1.0
python-docx>=1.1.2
openpyxl>=3.1.5
xlsxwriter>=3.2.0
markdown>=3.7
beautifulsoup4>=4.12.3
lxml>=5.3.0
scrapy>=2.12.0
selenium>=4.27.1
playwright>=1.49.1
streamlit>=1.40.2
gradio>=5.9.1
jupyter>=1.1.1
ipykernel>=6.29.5
notebook>=7.3.1
jupyterlab>=4.3.4
tensorboard>=2.18.0
wandb>=0.19.1
mlflow>=2.19.0
optuna>=4.1.0
hyperopt>=0.2.7
ray[tune]>=2.39.0
dask>=2024.12.1
joblib>=1.4.2
tqdm>=4.67.1
rich>=13.9.4
typer>=0.15.1
click>=8.1.8
loguru>=0.7.3
structlog>=24.4.0
prometheus_client>=0.21.0
psutil>=6.1.0
memory_profiler>=0.61.0
py-spy>=0.3.14
pytest>=8.3.4
pytest-asyncio>=0.24.0
pytest-mock>=3.14.0
pytest-cov>=6.0.0
black>=24.10.0
isort>=5.13.2
flake8>=7.1.1
mypy>=1.13.0
pre-commit>=4.0.1
bandit>=1.8.0
safety>=3.2.11
coverage>=7.6.9
locust>=2.32.4
gunicorn>=23.0.0
docker>=7.1.0
kubernetes>=31.0.0